+++ packages/attest_flutter/CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this package are documented here. The format follows
 [Keep a Changelog](https://keepachangelog.com/en/1.1.0/) and this project
 adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.6.1
+
+### Changed
+
+- Allow (and require) `attest` 0.6.0.
+
 ## 0.6.0
 
 ### Added